+++ sumatorio.h
@@ -1,5 +1,5 @@
 /* File busca.h */
 #ifndef SUMATORIO_H
 #define SUMATORIO_H
-float sumatorio ( float v[], int sizeDatos );
+double sumatorio ( double v[], int sizeDatos );
 #endif
